refactor : Bean 등록되지 않은 Class 생성시 예외 처리 추가
