+++ nextstep-di/src/main/java/nextstep/di/factory/InstantiateBeansException.java
@@ -1,0 +1,7 @@
+package nextstep.di.factory;
+
+public class InstantiateBeansException extends RuntimeException {
+    public InstantiateBeansException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
