+++ nextstep-di/src/main/java/nextstep/di/factory/NotRegisteredBeanException.java
@@ -1,0 +1,7 @@
+package nextstep.di.factory;
+
+public class NotRegisteredBeanException extends RuntimeException {
+    public NotRegisteredBeanException(String message) {
+        super(message);
+    }
+}
