+++ nextstep-di/src/test/java/nextstep/di/factory/example/service/TestService.java
@@ -1,9 +1,12 @@
-package nextstep.di.factory.example2;
+package nextstep.di.factory.example.service;
 
+import nextstep.annotation.Inject;
 import nextstep.stereotype.Service;
 
 @Service
 public class TestService {
+
+    @Inject
     public TestService(TestServiceObject testObject) {
     }
 }
