+++ nextstep-di/src/test/java/nextstep/di/factory/example/service/TestServiceObject.java
@@ -1,0 +1,4 @@
+package nextstep.di.factory.example.service;
+
+public class TestServiceObject {
+}
