--- nextstep-di/src/test/java/nextstep/di/factory/example2/TestServiceObject.java
@@ -1,4 +1,0 @@
-package nextstep.di.factory.example2;
-
-public class TestServiceObject {
-}
